Fixing typing issue
