+++ apps/admin/src/types/productAttributes.ts
@@ -1,0 +1,13 @@
+export interface ProductAttributes {
+    cpu_mfg: string;
+    cpu_type: string;
+    cpu_model: string;
+    cpu_assignment: string;
+    cpu_generation: number;
+    cpu_cores: number;
+    memory: number;
+    storage_size: number;
+    storage_type: string;
+    catagory: string;
+    size: string;
+}
